+++ ros2_ws/src/arm_motion_stack/arm_remote_control/src/remote_pc_heartbeat.cpp
@@ -58,8 +58,8 @@ public:
       });
 
     RCLCPP_INFO(
-      get_logger(), "Publishing remote-control heartbeat every %d ms on %s",
-      period_ms, topic.c_str());
+      get_logger(), "Publishing remote-control heartbeat every %lld ms on %s",
+      static_cast<long long>(period_ms), topic.c_str());
   }
 
 private:
